delete sub-examples
