+++ cggmp21/src/birkhoff/polynomial.rs
